Split MainVerticleIT into handler integration tests
