+++ src/test/java/edu/ucla/library/iiif/auth/handlers/AbstractHandlerIT.java
@@ -1,11 +1,10 @@
+package edu.ucla.library.iiif.auth.handlers;
 
-package edu.ucla.library.iiif.auth.verticles;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.api.extension.ExtendWith;
-
-import info.freelibrary.util.Logger;
 
 import edu.ucla.library.iiif.auth.Config;
 import edu.ucla.library.iiif.auth.services.DatabaseService;
@@ -14,15 +13,16 @@ import io.vertx.config.ConfigRetriever;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.client.WebClient;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
-import io.vertx.sqlclient.PoolOptions;
 
 /**
- * An abstract test that other tests can extend.
+ * A base class for handler integration tests.
  */
 @ExtendWith(VertxExtension.class)
-public abstract class AbstractHauthIT {
+@TestInstance(Lifecycle.PER_CLASS)
+public abstract class AbstractHandlerIT {
 
     /**
      * A test id.
@@ -35,9 +35,19 @@ public abstract class AbstractHauthIT {
     protected static final String TEST_ORIGIN = "https://iiif-test.library.ucla.edu";
 
     /**
-     * The configuration used to start the integration server.
+     * The application configuration.
      */
-    private JsonObject myConfig;
+    protected JsonObject myConfig;
+
+    /**
+     * A WebClient for calling the HTTP API.
+     */
+    protected WebClient myWebClient;
+
+    /**
+     * The port on which the application is listening.
+     */
+    protected int myPort;
 
     /**
      * Sets up the test.
@@ -45,12 +55,14 @@ public abstract class AbstractHauthIT {
      * @param aVertx A Vert.x instance
      * @param aContext A test context
      */
-    @BeforeEach
+    @BeforeAll
     public void setUp(final Vertx aVertx, final VertxTestContext aContext) {
         ConfigRetriever.create(aVertx).getConfig().compose(config -> {
             final DatabaseService db = DatabaseService.create(aVertx, config);
 
             myConfig = config;
+            myWebClient = WebClient.create(aVertx);
+            myPort = config.getInteger(Config.HTTP_PORT, 8888);
 
             // Add some database entries
             return CompositeFuture.all(db.setAccessLevel(TEST_ID, 0), db.setDegradedAllowed(TEST_ORIGIN, false))
@@ -64,34 +76,8 @@ public abstract class AbstractHauthIT {
      * @param aVertx A Vert.x instance
      * @param aContext A test context
      */
-    @AfterEach
+    @AfterAll
     public void tearDown(final Vertx aVertx, final VertxTestContext aContext) {
         aContext.completeNow();
     }
-
-    /**
-     * Returns the logger used by an extending test.
-     *
-     * @return The test's logger
-     */
-    protected abstract Logger getLogger();
-
-    /**
-     * Gets the pooling options for Vert.x's Postgres client.
-     *
-     * @return The pooling options for Vert.x's Postgres client
-     */
-    protected PoolOptions getPoolOpts() {
-        return new PoolOptions().setMaxSize(5);
-    }
-
-    /**
-     * Gets the port number for the test instance of hauth application.
-     *
-     * @return The application's port number
-     */
-    protected int getPort() {
-        return myConfig.getInteger(Config.HTTP_PORT, 8888);
-    }
-
 }
